src/test/java/_0701_0750/_713_Subarray_Product_Less_Than_K/SolutionITest.java
package _0701_0750._713_Subarray_Product_Less_Than_K;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertEquals;

/**
 * @author jacka
 * @version 1.0 on 5/17/2020
 */
class SolutionITest {

  @ParameterizedTest
  @MethodSource("solutionStream")
  void testOnlineCase1(Solution solution) {
    final int nums[] = {10, 5, 2, 6}, k = 100;
    assertEquals(8, solution.numSubarrayProductLessThanK(nums, k));
  }

  static Stream<Solution> solutionStream() {
    return Stream.of(
        new SolutionI()
    );
  }
}